+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 There are next changes:
 
+- use mb_orig_* functions
+- create directory /tmp/mocks for cache by default
+- constant SOFTMOCKS_ROOT_PATH marked as deprecated, use `\Badoo\SoftMocks::setProjectPath()` instead of it
+- fix create mocks cache dir race condition
+
 ## v2.0.0
 
 There are next changes: